Fixed sizeof() overflow error causeing segfault. sizeof on a c style array is the byte size, not element size!

+++ cmake_scripts/package.sh
@@ -13,5 +13,11 @@ cd build
 cpack
 cd ..
 
+if [ ! -f "./build/INTERCESSION-installer.exe" ]
+then
+    echo "Could not find installer executable, check cpack"
+    exit 1
+fi
+
+echo "****** Installer found! Moving to $PWD"
 mv build/INTERCESSION-installer.exe INTERCESSION-installer.exe
-echo "****** Moved installer to $PWD"
